refactor(scratch-org): move sharing of scratch org to ScratchOrgOperator
