+++ packages/sfp-cli/src/impl/pool/PoolFetchImpl.ts
@@ -6,7 +6,7 @@ import ScratchOrg from "@dxatscale/sfpowerscripts.core/lib/scratchorg/ScratchOrg
 import { getUserEmail } from "./services/fetchers/GetUserEmail";
 import ScratchOrgInfoFetcher from "./services/fetchers/ScratchOrgInfoFetcher";
 import ScratchOrgInfoAssigner from "./services/updaters/ScratchOrgInfoAssigner";
-import ShareScratchOrg from "@dxatscale/sfpowerscripts.core/lib/scratchorg/ShareScratchOrg";
+import ScratchOrgOperator from "@dxatscale/sfpowerscripts.core/lib/scratchorg/ScratchOrgOperator";
 import * as fs from "fs-extra";
 import SourceTrackingResourceController from "./SourceTrackingResourceController";
 import isValidSfdxAuthUrl from "./prequisitecheck/IsValidSfdxAuthUrl";
@@ -126,10 +126,9 @@ export default class PoolFetchImpl extends PoolBaseImpl {
       //Fetch the email for user id
       try {
         //Send an email for username
-        await new ShareScratchOrg(
+        await new ScratchOrgOperator(
           this.hubOrg,
-          soDetail
-        ).shareScratchOrgThroughEmail(emaiId);
+        ).shareScratchOrgThroughEmail(emaiId, soDetail);
       } catch (error) {
         SFPLogger.log(
           "Unable to send the scratchorg details to specified user. Check whether the user exists in the org",
